+++ test/system/data_source/test_cases/test_attribute_types.py
@@ -31,6 +31,10 @@ class TestAttributeTypes:
                 'bool_column': 'bool',
                 'list_column': 'str',
                 'runtime_column': 'bool',
+                'str_column_prov': 'str',
+                'int_column_prov': 'int',
+                'datetime_column_prov': 'datetime',
+                'bool_column_prov': 'bool',
             },
             'related': {
                 'str_column': 'str',
@@ -38,8 +42,18 @@ class TestAttributeTypes:
                 'datetime_column': 'datetime',
                 'bool_column': 'bool',
                 'list_column': 'str',
-                'root_int_column_min': 'double',
-                'root_int_column_max': 'double',
+                'root_int_column_min': 'float',
+                'root_int_column_max': 'float',
+                'root_str_column_min': 'str',
+                'root_str_column_max': 'str',
+                'root_datetime_column_min': 'datetime',
+                'root_datetime_column_max': 'datetime',
+                'root_int_column_prov_max': 'int',
+                'root_int_column_prov_min': 'int',
+                'root_str_column_prov_max': 'str',
+                'root_str_column_prov_min': 'str',
+                'root_datetime_column_prov_max': 'datetime',
+                'root_datetime_column_prov_min': 'datetime',
             }
         }
         assert data_source.attribute_types == expected
@@ -65,9 +79,21 @@ class TestAttributeTypes:
                     source_order=['source1', 'source2', 'source3', 'source4'],
                     return_type=None
                 ),
-                'str_column': ProvenanceField(
+                'int_column_prov': ProvenanceField(
                     source_order=['source1', 'source2', 'source3', 'source4'],
-                    return_type='keyword'
+                    return_type=None
+                ),
+                'str_column_prov': ProvenanceField(
+                    source_order=['source1', 'source2', 'source3', 'source4'],
+                    return_type=None
+                ),
+                'datetime_column_prov': ProvenanceField(
+                    source_order=['source1', 'source2', 'source3', 'source4'],
+                    return_type=None
+                ),
+                'bool_column_prov': ProvenanceField(
+                    source_order=['source1', 'source2', 'source3', 'source4'],
+                    return_type=None
                 ),
             },
         }